Added helocasting to SOG helicopters
