+++ addons/compat_sog/cfg/helicopters.hpp
@@ -1,0 +1,44 @@
+class vn_air_ch47_04_base;
+class vn_b_air_ch47_04_01: vn_air_ch47_04_base {
+    GVARMAIN(passengerTurrets)[] = {{1}, {2}};
+    GVARMAIN(rampAnims)[] = {{"ramp", 0, 1}};
+
+    EGVAR(staticLine,enabled) = 1;
+    EGVAR(staticLine,condition) = QUOTE(true);
+
+    EGVAR(helocast,boatPositions)[] = {CH47_HELOCAST_BOATPOS};
+    EGVAR(helocast,marker) = "Chemlight_green";
+};
+class vn_b_air_ch47_04_02: vn_air_ch47_04_base {
+    GVARMAIN(passengerTurrets)[] = {{1}, {2}};
+    GVARMAIN(rampAnims)[] = {{"ramp", 0, 1}};
+
+    EGVAR(staticLine,enabled) = 1;
+    EGVAR(staticLine,condition) = QUOTE(true);
+
+    EGVAR(helocast,boatPositions)[] = {CH47_HELOCAST_BOATPOS};
+    EGVAR(helocast,marker) = "Chemlight_green";
+};
+
+class vn_air_ch47_01_base;
+class vn_b_air_ch47_01_01: vn_air_ch47_01_base {
+    GVARMAIN(passengerTurrets)[] = {{1}, {2}};
+    GVARMAIN(rampAnims)[] = {{"ramp", 0, 1}};
+
+    EGVAR(staticLine,enabled) = 1;
+    EGVAR(staticLine,condition) = QUOTE(true);
+
+    EGVAR(helocast,boatPositions)[] = {CH47_HELOCAST_BOATPOS};
+    EGVAR(helocast,marker) = "Chemlight_green";
+};
+
+class vn_b_air_ch47_01_02: vn_air_ch47_01_base {
+    GVARMAIN(passengerTurrets)[] = {{1}, {2}};
+    GVARMAIN(rampAnims)[] = {{"ramp", 0, 1}};
+
+    EGVAR(staticLine,enabled) = 1;
+    EGVAR(staticLine,condition) = QUOTE(true);
+
+    EGVAR(helocast,boatPositions)[] = {CH47_HELOCAST_BOATPOS};
+    EGVAR(helocast,marker) = "Chemlight_green";
+};
