+++ addons/compat_sog/cfg/parachutes.hpp
@@ -1,0 +1,9 @@
+class vn_b_backpack_base;
+class vn_b_pack_ba18_01: vn_b_backpack_base {
+    GVARMAIN(isParachute) = 1;
+};
+
+class vn_o_backpack_base;
+class vn_o_pack_parachute_01: vn_o_backpack_base {
+    GVARMAIN(isParachute) = 1;
+};
