+++ addons/compat_sog/script_component.hpp
@@ -7,3 +7,4 @@
 // #define ENABLE_PERFORMANCE_COUNTERS
 
 #include "\z\aftb\addons\main\script_macros.hpp"
+#include "script_macros.hpp"
