+++ addons/compat_sog/script_macros.hpp
@@ -1,0 +1,1 @@
+#define CH47_HELOCAST_BOATPOS {0, 0, -1.5}
