+++ extra/example_subaddon/script_component.hpp
@@ -1,4 +1,4 @@
-#define SUBCOMPONENT subaddonName
+#define SUBCOMPONENT subaddonname
 #define SUBCOMPONENT_BEAUTIFIED Sub Addon
 
 #include "..\script_component.hpp"
